job-matching-system/src/scrapers/jobleads_scraper.py
"""
JobLeads scraper using requests and BeautifulSoup.
Note: This is a simplified implementation as JobLeads may have different structure.
"""
import time
import logging
import requests
from bs4 import BeautifulSoup
from typing import List, Dict, Any, Optional
from datetime import datetime, timedelta
import re
from urllib.parse import urljoin, quote_plus

logger = logging.getLogger(__name__)

class JobLeadsScraper:
    def __init__(self, user_agent: str = None):
        self.session = requests.Session()
        self.base_url = "https://jobleads.ca"  # Adjust based on actual JobLeads URL
        
        # Set up headers
        headers = {
            'User-Agent': user_agent or 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36',
            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8',
            'Accept-Language': 'en-US,en;q=0.5',
            'Accept-Encoding': 'gzip, deflate',
            'DNT': '1',
            'Connection': 'keep-alive',
            'Upgrade-Insecure-Requests': '1'
        }
        self.session.headers.update(headers)
    
    def search_jobs(self, search_terms: List[str], locations: List[str], 
                   days_back: int = 7) -> List[Dict[str, Any]]:
        """
        Search for jobs on JobLeads.
        
        Note: This is a placeholder implementation. 
        Actual implementation would need to be customized based on JobLeads' structure.
        """
        logger.warning("JobLeads scraper is a placeholder implementation")
        
        # Return empty list for now - would need actual JobLeads API/scraping logic
        return []
        
        # Placeholder code structure (commented out):
        """
        all_jobs = []
        
        for search_term in search_terms:
            for location in locations:
                try:
                    jobs = self._search_jobs_for_term(search_term, location, days_back)
                    all_jobs.extend(jobs)
                    time.sleep(2)  # Rate limiting
                    
                except Exception as e:
                    logger.error(f"Error searching for '{search_term}' in '{location}': {e}")
                    continue
        
        # Remove duplicates
        seen_urls = set()
        unique_jobs = []
        
        for job in all_jobs:
            job_url = job.get('application_link', '')
            if job_url not in seen_urls:
                seen_urls.add(job_url)
                unique_jobs.append(job)
        
        logger.info(f"Found {len(unique_jobs)} unique jobs on JobLeads")
        return unique_jobs
        """
    
    def _search_jobs_for_term(self, search_term: str, location: str, days_back: int) -> List[Dict[str, Any]]:
        """Search jobs for a specific term and location."""
        # Placeholder - would implement actual JobLeads search logic
        return []
    
    def _extract_job_data(self, element) -> Optional[Dict[str, Any]]:
        """Extract job data from a job element."""
        # Placeholder - would implement actual data extraction
        return None
    
    def _parse_posting_date(self, date_text: str) -> str:
        """Parse posting date text to standard format."""
        try:
            # Generic date parsing logic
            date_text = date_text.lower().strip()
            now = datetime.now()
            
            if 'today' in date_text or 'just posted' in date_text:
                return now.strftime('%Y-%m-%d')
            elif 'yesterday' in date_text:
                return (now - timedelta(days=1)).strftime('%Y-%m-%d')
            elif 'day' in date_text and 'ago' in date_text:
                days_match = re.search(r'(\d+)\s*day', date_text)
                if days_match:
                    days = int(days_match.group(1))
                    return (now - timedelta(days=days)).strftime('%Y-%m-%d')
            
            return now.strftime('%Y-%m-%d')
            
        except Exception:
            return datetime.now().strftime('%Y-%m-%d')
    
    def close(self):
        """Close the session."""
        self.session.close()
        logger.info("JobLeads scraper session closed")