job-matching-system/requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
playwright==1.40.0
openai==1.3.7
anthropic==0.7.8
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
pandas==2.1.4
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
lxml==4.9.3
python-docx==1.1.0
reportlab==4.0.7
PyPDF2==3.0.1
jinja2==3.1.2
openpyxl==3.1.2
smtplib-ssl==1.0.4